Backend/requirements.txt
gunicorn==23.0.0
Flask==3.1.0
requests
google-genai